Extract file system interaction to new file controller
